Using ES6 modules with underscore templates
Here I simplified the app code by having the template modules return the result of the template compile function - _.template always returns a callable object that when invoked will produce the template string. This cuts down on the amount of work required in the main module, just include the template and call the result with the parameters.

I also used the ES6 template string (not to be confused with templates as underscore knows them). This is basically a multi line string that cuts down on the amount of string building I need to do in code. The entire HTML template is now contained in a single string within the boilerplate template code.

You might be able to make a plugin that can get babel to understand a HTML file and automatically do the compile as if it was a JS file but that's more than I've done with babel.